Move sub-classes of "PropertyGridView" classes to their own files
